Decode returned JSON to PHP object
Added "stripe_decode" to config (and a blurb about what it does in the
readme) and condition applying json_decode if that config is set to
true. Set to false by default for anyone updating existing code (no
surprise breakage) and returns response with no decoding if the config
item does not exist for backwards compatibility.